Add a parameter for tile size
Tile sizes (in 3d world) wasn't controlled and we had different sizes depending on the tms detail level. i.e. Lvl16 was 611.5, lvl14 was 2446. And of course it gets huge numbers as the detail level decreases (millions for sub-10 detail levels). This commit will add a TileSize parameter under World.Settings and scale the base World object depending of that value. I'm not sure if scaling is the best solution (compared to generating smaller meshes) but it was easily the most practical and works for now.

Fixes #3